renamed files and reorganized app hierarchy for clearer separation of app and webclient code
