+++ README.md
@@ -6,7 +6,7 @@ Please ensure you have Node and MongoDB installed and configured.
 1. Clone Chandelier to your local environemnt
 2. Run npm install
 3. Start MongoDB. Create a new database named "chandelier"
-4. Run node app.js to start the API server
-5. Run gulp to start the web client
+4. cd \app and run node app.js to start the API server
+5. Run gulp default to start the web client
 
 Cheers, E
